# docs/openapi.json
{
  "swagger": "2.0",
  "info": {
    "title": "Lister API",
    "description": "API documentation for the Lister application",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/api/lists/{listId}/items/{itemId}/comments": {
      "get": {
        "description": "Get all comments for a specific item (not available for list owners)",
        "tags": [
          "comments"
        ],
        "parameters": [
          {
            "type": "string",
            "format": "uuid",
            "required": true,
            "in": "path",
            "name": "listId"
          },
          {
            "type": "string",
            "format": "uuid",
            "required": true,
            "in": "path",
            "name": "itemId"
          }
        ],
        "security": [
          {
            "basicAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "item": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "sourceUser": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "authorName": {
                    "type": "string"
                  },
                  "authorEmail": {
                    "type": "string",
                    "format": "email"
                  },
                  "text": {
                    "type": "string"
                  },
                  "createdAt": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "updatedAt": {
                    "type": "string",
                    "format": "date-time"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "post": {
        "description": "Create a new comment on an item (not available for list owners)",
        "tags": [
          "comments"
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "required": [
                "text"
              ],
              "properties": {
                "text": {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 5000
                }
              }
            }
          },
          {
            "type": "string",
            "format": "uuid",
            "required": true,
            "in": "path",
            "name": "listId"
          },
          {
            "type": "string",
            "format": "uuid",
            "required": true,
            "in": "path",
            "name": "itemId"
          }
        ],
        "security": [
          {
            "basicAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Default Response",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string",
                  "format": "uuid"
                },
                "item": {
                  "type": "string",
                  "format": "uuid"
                },
                "sourceUser": {
                  "type": "string",
                  "format": "uuid"
                },
                "authorName": {
                  "type": "string"
                },
                "authorEmail": {
                  "type": "string",
                  "format": "email"
                },
                "text": {
                  "type": "string"
                },
                "createdAt": {
                  "type": "string",
                  "format": "date-time"
                },
                "updatedAt": {
                  "type": "string",
                  "format": "date-time"
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/lists/{listId}/items/{itemId}/comments/{commentId}": {
      "get": {
        "description": "Get a specific comment by ID (not available for list owners)",
        "tags": [
          "comments"
        ],
        "parameters": [
          {
            "type": "string",
            "format": "uuid",
            "required": true,
            "in": "path",
            "name": "listId"
          },
          {
            "type": "string",
            "format": "uuid",
            "required": true,
            "in": "path",
            "name": "itemId"
          },
          {
            "type": "string",
            "format": "uuid",
            "required": true,
            "in": "path",
            "name": "commentId"
          }
        ],
        "security": [
          {
            "basicAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string",
                  "format": "uuid"
                },
                "item": {
                  "type": "string",
                  "format": "uuid"
                },
                "sourceUser": {
                  "type": "string",
                  "format": "uuid"
                },
                "authorName": {
                  "type": "string"
                },
                "authorEmail": {
                  "type": "string",
                  "format": "email"
                },
                "text": {
                  "type": "string"
                },
                "createdAt": {
                  "type": "string",
                  "format": "date-time"
                },
                "updatedAt": {
                  "type": "string",
                  "format": "date-time"
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "put": {
        "description": "Update an existing comment (not available for list owners)",
        "tags": [
          "comments"
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "text": {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 5000
                }
              }
            }
          },
          {
            "type": "string",
            "format": "uuid",
            "required": true,
            "in": "path",
            "name": "listId"
          },
          {
            "type": "string",
            "format": "uuid",
            "required": true,
            "in": "path",
            "name": "itemId"
          },
          {
            "type": "string",
            "format": "uuid",
            "required": true,
            "in": "path",
            "name": "commentId"
          }
        ],
        "security": [
          {
            "basicAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string",
                  "format": "uuid"
                },
                "item": {
                  "type": "string",
                  "format": "uuid"
                },
                "sourceUser": {
                  "type": "string",
                  "format": "uuid"
                },
                "authorName": {
                  "type": "string"
                },
                "authorEmail": {
                  "type": "string",
                  "format": "email"
                },
                "text": {
                  "type": "string"
                },
                "createdAt": {
                  "type": "string",
                  "format": "date-time"
                },
                "updatedAt": {
                  "type": "string",
                  "format": "date-time"
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "delete": {
        "description": "Delete a comment (not available for list owners)",
        "tags": [
          "comments"
        ],
        "parameters": [
          {
            "type": "string",
            "format": "uuid",
            "required": true,
            "in": "path",
            "name": "listId"
          },
          {
            "type": "string",
            "format": "uuid",
            "required": true,
            "in": "path",
            "name": "itemId"
          },
          {
            "type": "string",
            "format": "uuid",
            "required": true,
            "in": "path",
            "name": "commentId"
          }
        ],
        "security": [
          {
            "basicAuth": []
          }
        ],
        "responses": {
          "204": {
            "description": "Default Response",
            "schema": {}
          },
          "401": {
            "description": "Default Response",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "host": "localhost:3000",
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "securityDefinitions": {
    "basicAuth": {
      "type": "basic"
    }
  },
...
}